db 비교 기능 수정
1. db 비교 역비교 , 대소문자 비교 기능 추가
2. postgresql 버그 수정

+++ varsql-core/src/main/java/com/varsql/core/db/datatype/handler/MetaDataHandler.java
@@ -1,7 +1,7 @@
 package com.varsql.core.db.datatype.handler;
 
 public interface MetaDataHandler {
-	default int getColumnLength(int length){
+	default long getColumnLength(long length){
 		return length;
 	}
 }
